refactor(storage): restructure storage package into focused modules
Resolves #23: Refactor large files for improved maintainability

This commit restructures the monolithic storage/store.go (~1600 lines) into
smaller, focused components following the single responsibility principle:

- card_store.go: Core CardStore struct and initialization
- card_ops.go: Card CRUD operations
- deck_ops.go: Deck directory management
- review.go: Spaced repetition functionality
- stats.go: Statistics and analytics
- search.go: Card searching and filtering
- io/: File system operations
- parser/: Markdown parsing and formatting

Benefits:
- No file exceeds 300 lines, making code more maintainable
- Clear separation of concerns
- Improved code organization
- Better testability of individual components
- Groundwork for issue #14 (file watching) implementation

All tests pass with the new structure, maintaining compatibility with UI layer.
