+++ internal/storage/io/io_test.go
@@ -1,0 +1,10 @@
+package io
+
+import (
+	"testing"
+)
+
+// TestIOFunctions verifies io functionality
+func TestIOFunctions(t *testing.T) {
+	// This is a minimal test to ensure the package can be imported
+}
