+++ internal/storage/parser/parser_test.go
@@ -1,0 +1,10 @@
+package parser
+
+import (
+	"testing"
+)
+
+// TestParserFunctions verifies parser functionality
+func TestParserFunctions(t *testing.T) {
+	// This is a minimal test to ensure the package can be imported
+}
