+++ internal/ui/ui_test.go
@@ -1,0 +1,16 @@
+package ui
+
+import (
+	"testing"
+
+	"github.com/DavidMiserak/GoCard/internal/storage"
+)
+
+// TestUIImports is a simple test to verify imports
+func TestUIImports(t *testing.T) {
+	// Initialize a simple test to check which methods are needed
+	var store storage.CardStore
+
+	// Just to prevent unused variable warnings
+	_ = store
+}
