+++ internal/ui/views/views_test.go
@@ -1,0 +1,14 @@
+package views
+
+import (
+	"testing"
+
+	"github.com/DavidMiserak/GoCard/internal/storage"
+)
+
+// This is a minimal test file to verify imports are working
+
+func TestViewImports(t *testing.T) {
+	// Just make sure the imports are correct, we don't need to run anything
+	var _ storage.CardStore
+}
